reflection to check if type implements INodeViewCustomization don't import assembly if its marked node lib and contains any nodemodels and any nodecustomizations this means we wont import ZT nodes from assemblies that contain customziations and are marked node libraries.
remove some unusued stuff from package man view ext

move generic check to type extensions utils
