+++ src/DynamoCore/Core/NodeModelAssemblyLoader.cs
@@ -7,6 +7,7 @@ using Dynamo.Configuration;
 using Dynamo.Graph.Nodes;
 using Dynamo.Logging;
 using Dynamo.Migration;
+using Dynamo.Utilities;
 
 namespace Dynamo.Models
 {
@@ -197,6 +198,22 @@ namespace Dynamo.Models
         internal static bool ContainsNodeModelSubType(Assembly assem)
         {
             return assem.GetTypes().Any(IsNodeSubType);
+        }
+        internal static bool ContainsNodeViewCustomizationType(Assembly assem)
+        {
+            return GetCustomizationTypesUsingReflection(assem).Any();
+        }
+
+        internal static IEnumerable<Type> GetCustomizationTypesUsingReflection(Assembly assem)
+        {
+
+            var customizerType = Type.GetType("Dynamo.Wpf.INodeViewCustomization`1,DynamoCoreWpf");
+            if(customizerType != null)
+            {
+                var customizerImps = assem.GetTypes().Where(t => !t.IsAbstract && TypeExtensions.ImplementsGeneric(customizerType, t));
+                return customizerImps;
+            }
+            return new Type[] { };
         }
 
         /// <summary>
