+++ src/DynamoCoreWpf/NodeViewCustomization/NodeViewCustomizationLoader.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Dynamo.Logging;
+using Dynamo.Utilities;
 using Dynamo.Wpf.Properties;
 
 namespace Dynamo.Wpf
@@ -50,7 +51,8 @@ namespace Dynamo.Wpf
         private static IEnumerable<Type> GetCustomizationTypes(Assembly assem)
         {
             var customizerType = typeof(INodeViewCustomization<>);
-            var customizerImps = assem.GetTypes().Where(t => !t.IsAbstract && ImplementsGeneric(customizerType, t));
+            var customizerImps = assem.GetTypes().Where(t => !t.IsAbstract && 
+                TypeExtensions.ImplementsGeneric(customizerType, t));
             return customizerImps;
         }
 
@@ -106,22 +108,7 @@ namespace Dynamo.Wpf
             //return types.First().TypeHierarchy().First(t => counts[t] == total);
         }
 
-        private static bool ImplementsGeneric(Type generic, Type toCheck)
-        {
-            var cur = toCheck.IsGenericType ? toCheck.GetGenericTypeDefinition() : toCheck;
-
-            var isGenInterf = cur.GetInterfaces().Any(
-                x =>
-                    x.IsGenericType &&
-                        x.GetGenericTypeDefinition() == generic);
-
-            if (generic == cur || isGenInterf)
-            {
-                return true;
-            }
-
-            return false;
-        }
+       
 
         internal static bool ContainsNodeViewCustomizationType(Assembly assem)
         {
