+++ src/DynamoPackagesWPF/PackageManagerViewExtension.cs
@@ -15,7 +15,6 @@ namespace Dynamo.PackageManager.UI
     {
         private readonly List<IViewExtension> requestedExtensions = new List<IViewExtension>();
         private PackageManagerExtension packageManager;
-        private ExtensionLibraryLoader librayLoader;
         public string Name
         {
             get
@@ -50,10 +49,10 @@ namespace Dynamo.PackageManager.UI
 
         public void Loaded(ViewLoadedParams viewLoadedParams)
         {
-            InspectPackageForNodeViewCustomizationBinaries(packageManager?.PackageLoader.LocalPackages);
+            // Do nothing for now
         }
 
-        public void Shutdown()
+            public void Shutdown()
         {
             // Do nothing for now
         }
@@ -73,33 +72,9 @@ namespace Dynamo.PackageManager.UI
                 packageManager.PackageLoader.PackgeLoaded += packageLoadedHandler;
                 var packagesToCheck = packageManager.PackageLoader.LocalPackages;
                 RequestLoadViewExtensionsForLoadedPackages(packagesToCheck);
-                // also look for assemblies that only contains view customization (no nodeModels) as they won't have
-                // been imported into the customizationLibrary previously.
-                InspectPackageForNodeViewCustomizationBinaries(packagesToCheck);
             }
         }
 
-        private void InspectPackageForNodeViewCustomizationBinaries(IEnumerable<Package> packages)
-        {
-            foreach (var package in packages)
-            {
-                // these packages are already loaded, these assemblies which contain only INodeViewCustomizatons
-                // should already be loaded as well - so we need to check packageAssemblies that are marked as NodeLibraies
-                // check if they contain INodeViewCustomization and if so inject them into the CustmizationLibrary
-                if (package.Loaded)
-                {
-                    foreach(var nodeLibAssem in package.LoadedAssemblies.Where(x => x.IsNodeLibrary))
-                    {
-                        // if this assembly is a nodeLibrary(was in nodelibraries list) and it contains customizations
-                        // TODO be careful of loading customizations twice.
-                        if (Wpf.NodeViewCustomizationLoader.ContainsNodeViewCustomizationType(nodeLibAssem.Assembly))
-                        {
-                            librayLoader.LoadNodeViewCustomizationAssembly(nodeLibAssem.Assembly);
-                        }
-                    }
-                }
-            }
-        }
 
         private void RequestLoadViewExtensionsForLoadedPackages(IEnumerable<Package> packages)
         {
